requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain
chromadb
tiktoken
langchain-google-genai
langchain-community
python-dotenv
google-generativeai 